+++ CHANGELOG.md
@@ -4,10 +4,16 @@ All notable changes to this add-on will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
+
+
+## [11] - 2021-09-07
 ### Added
 - active/Cross Site WebSocket Hijacking.js > an active scan for Cross-Site WebSocket Hijacking vulnerability
 - targeted/cve-2021-22214.js > A targeted script to check for Unauthorised SSRF on GitLab - CVE 2021-22214
 - httpsender/full-session-n-csrf-nashorn.js > full session and csrf token management.
+- httpfuzzerprocessor/unexpected_responses.js > compare response codes to a (pass/fail) regex and generate alerts
+- targeted/dns-email-spoofing > Check if DMARC / SPF policies are configured on a domain.
+- httpsender/add-more-headers.js > Add caller-specified headers to all requests.
 - authentication/OfflineTokenRefresh.js - refresh oauth2 offline tokens
 - httpsender/AddBearerTokenHeader.js - refresh oauth2 offline tokens
 
@@ -129,6 +135,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - First packaged version
 
-[Unreleased]: https://github.com/zaproxy/community-scripts/compare/v10...HEAD
+[Unreleased]: https://github.com/zaproxy/community-scripts/compare/v11...HEAD
+[11]: https://github.com/zaproxy/community-scripts/compare/v10...v11
 [10]: https://github.com/zaproxy/community-scripts/compare/v9...v10
 [9]: https://github.com/zaproxy/community-scripts/compare/7278617af4bd1bc3d8db41c00a0437bf78ba6e51...v9